separate transform data model
