+++ bofire/data_models/transforms/api.py
@@ -1,0 +1,3 @@
+from bofire.data_models.transforms.remove import RemoveTransform
+
+AnyTransform = RemoveTransform
